requirements.txt
alembic==0.9.10
aniso8601==3.0.2
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
autoenv==1.0.0
bleach==2.1.2
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.1.2
defer==1.0.6
distro-info==0.18
entrypoints==0.2.3.post1
Flask==0.12.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.10.0
Flask-Migrate==2.2.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
linecache2==1.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
oauth==1.0.1
olefile==0.45.1
pandocfilters==1.4.2
passlib==1.7.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.7.1
prompt-toolkit==1.0.15
protobuf==3.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.5.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.4.2
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
pytest==3.7.1
python-apt==1.6.2
python-dateutil==2.6.1
python-debian==0.1.32
python-editor==1.0.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.9
stevedore==1.29.0
system-service==0.3
systemd-python==234
terminado==0.7
testpath==0.3.1
testresources==2.0.0
tornado==4.5.3
traceback2==1.4.0
traitlets==4.3.2
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
widgetsnbextension==2.0.0
wrapt==1.10.11
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
